pyproject.toml
[project]
name = "sam_audio"
version = "0.1.0"
description = "Segment Anything Audio"
authors = [
    { name="Andros Tjandra", email="androstj@meta.com" },
    { name="Ann Lee", email="annl@meta.com" },
    { name="Bowen Shi", email="bshi@meta.com" },
    { name="Julius Richter", email="jrichter@meta.com" },
    { name="Matt Le", email="mattle@meta.com" },
    { name="Yi-Chiao Wu", email="yichiaowu@meta.com" },
]

readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "dacvae@git+https://github.com/facebookresearch/dacvae.git",
    "audiobox_aesthetics",
    "einops",
    "imagebind@git+https://github.com/facebookresearch/ImageBind.git",
    "laion-clap",
    "numpy",
    "perception-models@git+https://github.com/facebookresearch/perception_models@unpin-deps",
    "pydub",
    "torch",
    "torchaudio",
    "torchcodec",
    "torchdiffeq",
    "torchvision",
    "transformers>=4.54.0",
]

[tool.setuptools.packages.find]
include = ["sam_audio"]

[tool.ruff]

target-version = "py310"

lint.select=[
    "B",
    "C",
    "E",
    "W",
    "F",
    "I",
]
lint.ignore = [
    "E501",
    "E731",
    "C901",
    "B006",
]

[project.urls]
Homepage = "https://github.com/facebookresearch/sam-audio"
Repository = "https://github.com/facebookresearch/sam-audio"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
